backup/requirements.txt
firebase-admin>=6.2.0
python-dotenv>=1.0.0